Refactor nextBoardState() for accuracy
Changed the conditions for counting a neighbor alive to comply with the rules intended for the game (i.e. not considering an adjacent cell on the other side of the board as a neighbor). This more accurately reflects the intended rules.

Used a simpliefed way of going through each neighbor using an array for directions.